% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-data-wrangling.R
\name{add.count}
\alias{add.count}
\title{Add Count}
\usage{
add.count(df, rename = "N")
}
\arguments{
\item{df}{data.frame object}

\item{rename}{Specify to rename column for counts}
}
\value{
data.frame object
}
\description{
Typically used after group_by(). This function gets the counts for each group and adds
it back to the ungrouped dataset.
}
\examples{
library(tidyverse)
mtcars\%>\%
  group_by(am) \%>\%
  add.count()

}
