% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_sparkline.R
\name{appendColumnDefs}
\alias{appendColumnDefs}
\title{Append a column definition to the column definitions in options}
\usage{
appendColumnDefs(options, def)
}
\arguments{
\item{options}{list of DT options}

\item{def}{list representing column definition}
}
\value{
options appended with new definition
}
\description{
Append a column definition to the column definitions in options
}
