% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_sparkline.R
\name{appendfnDrawCallback}
\alias{appendfnDrawCallback}
\title{Append a fnDrawCallback row into the existing js}
\usage{
appendfnDrawCallback(js, template)
}
\arguments{
\item{js}{The existing js}

\item{template}{the js to be inserted into the fdDrawCallback}
}
\value{
js containing the fnDrawCallback with the template inserted
}
\description{
Append a fnDrawCallback row into the existing js
}
