% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-collapse-recode-variables.R
\name{collapse.var}
\alias{collapse.var}
\title{Collapse Multiple Variables into a single Column}
\usage{
collapse.var(df, var = colnames(df), return.conflicts = FALSE)
}
\arguments{
\item{df}{data.frame object}

\item{var}{Variable to collapse}

\item{return.conflicts}{If set to TRUE and there are conflicts within the input variables, the conflicting rows will be returned}
}
\value{
vector object. If return.conflicts set to TRUE, a list object is returned
}
\description{
Collapse Multiple Variables into a single Column
}
\examples{
df <- data.frame(
v1.t1 = c(1:10),
v1.t2 = c(1:5, rep(NA, 5)),
v2.t1 = c(12,13,14,rep(NA,7)),
v2.t2 = c(11:20)
)

df$v1 <- collapse.var(df, var = c("v1.t1","v1.t2"))
df$v2 <- collapse.var(df, var = c("v2.t1","v2.t2"))
conflicts <- collapse.var(df, var = c("v2.t1","v2.t2"), return.conflicts = T)
 
also see <https://msrcodelibrary.netlify.app/2020/06/10/multiple-recoding-and-collapsing-columns/> for 
step-by-step tutorial.

}
