% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_sparkline.R
\name{colSparkline}
\alias{colSparkline}
\title{create a list to be used as a column definition in a DT datatable}
\usage{
colSparkline(columns, colnames, rownames, class_suffix = "", width = "5\%",
  render_js = NULL, ...)
}
\arguments{
\item{columns}{column names or indicies}

\item{colnames}{column names in data set}

\item{rownames}{row names in data set}

\item{class_suffix}{class name will be spark + suffix}

\item{width}{Width of column}

\item{render_js}{alternative js to render}

\item{...}{additional options passed to the column definition}
}
\description{
create a list to be used as a column definition in a DT datatable
}
