% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-computation-functions.R
\name{cor.with}
\alias{cor.with}
\title{Correlation with Grouping Variable}
\usage{
cor.with(df, var, with = colnames(df), group = NULL, dp = 2)
}
\arguments{
\item{df}{data.frame object}

\item{var}{Variables of main interest}

\item{with}{Variable to correlate with. Defaults to all other in df}

\item{group}{Grouping variable. If specified, correlation will be performed within each group separately}

\item{dp}{Percentage of missing data will be reported. dp allows re-specification of number of decimal places}

\item{id}{Identifier variable. Input as string}
}
\value{
data.frame object
}
\description{
Perform correlations with the main variables of interest.
Group can be specified to perform correlations within different subsets
of the data.
}
\examples{
library(car)
cor.with(mtcars, var = c("vs"), group = "am")
cor.with(mtcars, var = c("vs", "cyl"))

}
