% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dt_list_of_lags}
\alias{dt_list_of_lags}
\title{to be used within data.table data to create list columns (use := list(dt_list_of_lags()))}
\usage{
dt_list_of_lags(x_vector, lags_x)
}
\arguments{
\item{x_vector}{column to be lagged}

\item{lags_x}{vector of lag values (eg c(0:2))}
}
\value{
listcolumn of lagged vectors
}
\description{
to be used within data.table data to create list columns (use := list(dt_list_of_lags()))
}
