% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dt_list_of_lags_2}
\alias{dt_list_of_lags_2}
\title{to be used within data.table data to create list columns (use := list(dt_list_of_lags()))}
\usage{
dt_list_of_lags_2(x_vector_1, x_vector_2, lags_x)
}
\arguments{
\item{x_vector_1}{column to be lagged}

\item{x_vector_2}{column to be lagged}

\item{lags_x}{vector of lag values (eg c(0:2))}
}
\value{
listcolumn of lagged vectors with : separator
}
\description{
to be used within data.table data to create list columns (use := list(dt_list_of_lags()))
}
