% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-handling-missing-data.R
\name{estimate.listwise.deletion}
\alias{estimate.listwise.deletion}
\title{Estimate Listwise Deletion}
\usage{
estimate.listwise.deletion(
  df,
  var = colnames(df),
  dp = 2,
  return.num = FALSE,
  return.perc = FALSE,
  print.text = TRUE
)
}
\arguments{
\item{df}{data.frame object}

\item{var}{Input column names (defaults to using all columns in df)}

\item{dp}{Number of decimal places for percentage}

\item{return.num}{If set to TRUE, returns number of missing cases}

\item{return.perc}{If set to TRUE, returns percentage of missing cases}

\item{print.text}{If set to FALSE, missing cases are not printed}
}
\value{
Default none. When return.num or return.perc equals TRUE, numeric object is returned
}
\description{
Estimate number of missingness caused by listwise deletion
}
