% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-reporting.R
\name{extract.lm}
\alias{extract.lm}
\title{Extract statistics from lm object}
\usage{
\method{extract}{lm}(
  lm.model,
  var = c("Estimate", "Std. Error", "t value", "Pr(>|t|)"),
  include.intercept = TRUE
)
}
\arguments{
\item{lm.model}{lm object}

\item{var}{Variable. Set to "Estimate", "Std. Error", "t value", or "Pr(>|t|)"}

\item{include.intercept}{If set to FALSE, ignores estimates from intercept}
}
\value{
A vector
}
\description{
gets summary() of lm object and returns specified statistics from coefficient table
}
\examples{
lm.model <- lm(y ~ x)
extract.lm(lm.model, var = "Estimate")
extract.lm(lm.model, var = "Std. Error")
extract.lm(lm.model, var = "Pr(>|t|)")
}
