% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_callback.R
\name{formatArrow}
\alias{formatArrow}
\title{Add trend arrows to a DT datatable}
\usage{
formatArrow(table, columns, valueColumns = columns, arrowHTML = "&#8680;",
  titleColumn = NULL)
}
\arguments{
\item{table}{the DT datatable}

\item{columns}{column names or indicies to be replaced with arrows}

\item{valueColumns}{values to use to rotate the arrows (defaults to columns)}

\item{arrowHTML}{arrow symbol to use (defaults to &#8680;)}

\item{titleColumn}{column to use as a tooltip (defaults to NULL)}
}
\value{
table updated with arrow formatting
}
\description{
Add trend arrows to a DT datatable
}
