% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_sparkline.R
\name{formatSparkline}
\alias{formatSparkline}
\title{Add a sparkline column to a DT datatable}
\usage{
formatSparkline(table, columns, sparklineOpts, class_suffix = "",
  width = "5\%", ...)
}
\arguments{
\item{table}{The DT}

\item{columns}{column names or indicies of the data to be used as sparkline columns}

\item{sparklineOpts}{options passed to sparkline - use spark_options to create}

\item{class_suffix}{optional suffix (prefixed with spark) - randomly generated if not provided}

\item{width}{column width}

\item{...}{Other options passed to the column definition}
}
\value{
table DT with updated options and dependencies
}
\description{
Add a sparkline column to a DT datatable
}
