% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_callback.R
\name{formatTooltip}
\alias{formatTooltip}
\title{Add trend arrows to a DT datatable}
\usage{
formatTooltip(table, columns, valueColumns = columns, titleColumn)
}
\arguments{
\item{table}{the DT datatable}

\item{columns}{column names or indicies to be replaced with arrows}

\item{valueColumns}{values to use to rotate the arrows (defaults to columns)}

\item{titleColumn}{column to use as a tooltip (defaults to NULL)}
}
\value{
table updated with arrow formatting
}
\description{
Add trend arrows to a DT datatable
}
