% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{least_squares_over_lags}
\alias{least_squares_over_lags}
\title{Calcaulte least squares gradiant}
\usage{
least_squares_over_lags(x, lags_x, re_scale_x = 1)
}
\arguments{
\item{x}{column to be lagged}

\item{lags_x}{vector of lag values (eg c(0:2))}

\item{re_scale_x}{x if required}
}
\value{
vector of trend values
}
\description{
Calcaulte least squares gradiant
}
