% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list_to_name_value_chr}
\alias{list_to_name_value_chr}
\title{convert a list to a string of name:values}
\usage{
list_to_name_value_chr(l)
}
\arguments{
\item{l}{a list}
}
\value{
string of name:values from list
}
\description{
convert a list to a string of name:values
}
