% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{logical_to_js}
\alias{logical_to_js}
\title{convert logical values to javascript true/false}
\usage{
logical_to_js(x)
}
\arguments{
\item{x}{vector to convert}
}
\value{
character vector
}
\description{
convert logical values to javascript true/false
}
