% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-computation-functions.R
\name{meta.with}
\alias{meta.with}
\title{Meta-analysis with Grouping Variable}
\usage{
meta.with(
  df,
  estimate = "estimate",
  se = "se",
  group = NULL,
  random = T,
  na.rm = T
)
}
\arguments{
\item{df}{data.frame object}

\item{estimate}{Regression estimates}

\item{se}{Regression standard errors}

\item{group}{Grouping variable}

\item{random}{If set to TRUE, random-effects meta-analysis will be run. If set to FALSE, fixed-effects meta-analysis will be run}

\item{na.rm}{If set to TRUE, NAs will be ignored in computation of weights}
}
\value{
data.frame object
}
\description{
Meta-analysis with Grouping Variable
}
\examples{
df <- data.frame(
      estimate = c(1,2,3,2,1,4),
      se = c(.1,.5,.3,.8,1,.7),
      group = c("a","a","a","b","b","b")
      )
meta.with(df, group = "group")

}
