% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-data-wrangling.R
\name{multi.spread}
\alias{multi.spread}
\title{Spread Multiple Variables}
\usage{
multi.spread(df, id, key, var = NULL, sep = "_", custom.append = NULL)
}
\arguments{
\item{df}{data.frame object}

\item{id}{Identifier variable. Input as string}

\item{key}{Key variable to spread on. Input as string}

\item{var}{Variables being spread. Defaults to all other variables (apart from id and key)}

\item{sep}{Separator that appends to variable names after being spread}

\item{custom.append}{Customised labels that are appended to variable names. Length should match number of unique keys}
}
\value{
data.frame object
}
\description{
Wrangles data from long to wide form
}
\examples{
df <- data.frame(
      id = c(1,1,1,2,2,2,3,3,3),
      time = c(1,2,3,1,2,3,1,2,3),
      v1 = 1:9,
      v2 = 11:19,
      v3 = paste0("a",1:9)
      )
df.wide <- multi.spread(df, id = "id", key = "time") 
df.wide <- multi.spread(df, id = "id", key = "time", custom.append = c("t1","t2","t3")) 
df.wide <- multi.spread(df, id = "id", key = "time", var = c("v1","v2")) 
}
