% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-simulation.R
\name{plot.convergence}
\alias{plot.convergence}
\title{Plot Convergence for Simulations}
\usage{
\method{plot}{convergence}(x)
}
\arguments{
\item{x}{vector object. Index should match number of iterations}
}
\value{
ggplot object. Plots mean of estimate against number of iterations
}
\description{
Over each iteration, the function will compute the mean estimate
of the simulated statistic for all past iterations.
Curve should stabilise if convergence is reached.
}
\examples{
set.seed(100)
flip.coin <- function(n){
  return(sample(0:1, size=n, replace = T))
}
flips <- flip.coin(100)
plot.convergence(flips)

}
