% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-simulation.R
\name{plot.sampler}
\alias{plot.sampler}
\title{Plot Density Plot given a samples from a sampling function}
\usage{
\method{plot}{sampler}(sampler)
}
\arguments{
\item{sampler}{vector object}
}
\value{
density plot
}
\description{
Plot Density Plot given a samples from a sampling function
}
\examples{
set.seed(100)
plot.sampler(rnorm(10000))
plot.sampler(runif(1000, min = 0, max = 1))

}
