% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-computation-functions.R
\name{prepare.roc}
\alias{prepare.roc}
\title{Prepare ROC}
\usage{
prepare.roc(df, positive, negative)
}
\arguments{
\item{df}{data.frame object}

\item{positive}{Variables which has the "correct" answer of 1}

\item{negative}{Variables which has the "correct" answer of 0}
}
\value{
data.frame object
}
\description{
Computes number of true positives, false positive, true negatives,
and false negatives for each participant
}
\examples{
set.seed(100)
df <- data.frame(
     p1 = sample(0:1, size = 10, replace = T),
     p2 = sample(0:1, size = 10, replace = T),
     n1 = sample(0:1, size = 10, replace = T),
     n2 = sample(0:1, size = 10, replace = T)
     )
prepare.roc(df, positive = c("p1","p2"), negative = c("n1","n2"))

}
