% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-collapse-recode-variables.R
\name{recode.multiple}
\alias{recode.multiple}
\title{Recode Multiple Variables}
\usage{
\method{recode}{multiple}(df, var, from, to = 1:length(from), fun = as.numeric)
}
\arguments{
\item{df}{data.frame object}

\item{var}{Variable to recode}

\item{from}{Original scale (order of input needs to match desired scale)}

\item{to}{Desired scale (order of input needs to match original scale). Defaults to integer scale starting from 1}

\item{fun}{Function determining object type of output variables}
}
\value{
data.frame object
}
\description{
Recodes variables that share the same scale
}
\examples{
df <- data.frame(
      v1 = c("Strongly Disagree","Disagree", "Neutral"," Agree", "Strongly Agree"),
      v2 = c("Strongly Disagree","Disagree","Strongly Disagree","Disagree","Strongly Disagree"),
      v3 = c( "Neutral"," Agree", "Strongly Agree", "Neutral"," Agree")
      )

recode.multiple(df, var = c("v1","v2","v3"), from = c("Strongly Disagree","Disagree", "Neutral"," Agree", "Strongly Agree"))
recode.multiple(df, var = c("v1","v2","v3"), from = c("Strongly Disagree","Disagree", "Neutral"," Agree", "Strongly Agree"), to = c(10,20,30,40,50))
recode.multiple(df, var = c("v1","v2","v3"), from = c("Strongly Disagree","Disagree", "Neutral"," Agree", "Strongly Agree"), 
                to = c("ten", "twenty", "thirty", "forty", "fifty"),
                fun = function(x){factor(x, levels = c("ten", "twenty", "thirty", "forty", "fifty"))})

}
