% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-computation-functions.R
\name{regress.with}
\alias{regress.with}
\title{Regression with Grouping Variable}
\usage{
regress.with(
  form,
  df,
  group,
  output = c("Estimate", "Std. Error", "t value", "Pr(>|t|)"),
  dp = 2
)
}
\arguments{
\item{form}{Input formulae as string}

\item{df}{data.frame object}

\item{group}{Grouping variable. If specified, correlation will be performed within each group separately}

\item{output}{Desired statistical output. Set to "Estimate", "Std. Error", "t value", or "Pr(>|t|)"}

\item{dp}{Percentage of missing data will be reported. dp allows re-specification of number of decimal places}
}
\value{
data.frame object
}
\description{
Perform separate regressions for each unique group
}
\examples{
library(car)
regress.with("mpg ~ wt + cyl",group="am", mtcars, output = "Estimate")
regress.with("mpg ~ wt + cyl",group="am", mtcars, output = 2) # standard error

}
