% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-reporting.R
\name{report.ci}
\alias{report.ci}
\title{Reporting confidence intervals}
\usage{
report.ci(est, se, alpha = 0.05, dp = 2, start = NULL, end = NULL)
}
\arguments{
\item{est}{Estimate of Population Mean}

\item{se}{Standard Error}

\item{alpha}{Type I error. Two-tailed test is assumed}

\item{dp}{Number of decimal places}

\item{start}{Character string appended to start of output}

\item{end}{Character string appended to end of output}
}
\value{
A character string reporting the lower and upper limits of CI
}
\description{
Reporting confidence intervals
}
\examples{
report.ci(1, .5);
report.ci(1, .5, start = "95\% CI:");
}
