% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-reporting.R
\name{report.p}
\alias{report.p}
\title{Reporting p values}
\usage{
report.p(p, dp = 3, start = NULL, end = NULL)
}
\arguments{
\item{p}{Input p value}

\item{dp}{Number of decimal places}

\item{start}{Character string appended to start of output}

\item{end}{Character string appended to end of output}
}
\value{
A character string with '= ' (or '< .001' if p falls below threshold) before the input value
}
\description{
Reporting p values
}
\examples{
report.p(.02);
report.p(.00003, start = "p ");
}
