% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-reporting.R
\name{report}
\alias{report}
\title{Report}
\usage{
report(x, dp = 2)
}
\arguments{
\item{x}{Input vector object}

\item{dp}{Number of decimal places}
}
\value{
vector object
}
\description{
Short-hand for round function
}
