% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-simulation.R
\name{resample}
\alias{resample}
\title{Resample}
\usage{
resample(x, n)
}
\arguments{
\item{x}{vector object}

\item{n}{Number of samples to draw}
}
\value{
vector object
}
\description{
Resamples from a given vector. Useful for bootstrapping
}
\examples{
set.seed(100)
x <- rnorm(50)
resample(x, 50)
}
