% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-computation-functions.R
\name{rmse}
\alias{rmse}
\title{Compute Root Mean Square Error}
\usage{
rmse(x, na.rm = T)
}
\arguments{
\item{x}{Input vector object}

\item{na.rm}{If TRUE, NAs are ignored}
}
\value{
A vector object
}
\description{
Compute Root Mean Square Error
}
