% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-data-wrangling.R
\name{rowbind.match}
\alias{rowbind.match}
\title{Row Bind Matching Column Names}
\usage{
rowbind.match(df.old, x.new, names = colnames(df.old))
}
\arguments{
\item{df.old}{data.frame object}

\item{x.new}{vector object. Should be named carefully}

\item{names}{Desired column names for output df}
}
\value{
data.frame object
}
\description{
row binds a vector to an existing data.frame object (data.frame can be empty).
The vector should be named.
This function ensures that the vector values are added to the column
that shares their respectives names.
}
\examples{
df <- data.frame()
x <- 1:5
names(x) <- c("a","b","c","f","g")
rbind.match(df, x, names = c("a","b","c","d","e")) 
rbind.match(df, x, names = c("g","f","c","b","a")) 

}
