% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-computation-functions.R
\name{rowSD}
\alias{rowSD}
\title{rowSD similar to rowMeans and rowSums}
\usage{
rowSD(df, na.rm = T)
}
\arguments{
\item{df}{data.frame object}

\item{na.rm}{If set to TRUE, NAs are ignore in computation}
}
\value{
data.frame object
}
\description{
rowSD similar to rowMeans and rowSums
}
\examples{
set.seed(100) 
df <- data.frame(
      a = rnorm(10),
      b = rnorm(10),
      c = rnorm(10)
      )
rowSD(df)

}
