% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_helper.R
\name{spark_options}
\alias{spark_options}
\title{Create sparkline options js to include in column call back}
\usage{
spark_options(type = NULL, width = NULL, height = NULL,
  lineColor = NULL, fillColor = NULL, chartRangeMin = NULL,
  chartRangeMax = NULL, composite = NULL, enableTagOptions = NULL,
  tagOptionPrefix = NULL, tagValuesAtribute = NULL,
  disableHiddenCheck = NULL, spark_options_chart_type = NULL, ...)
}
\arguments{
\item{type}{One of 'line' (default), 'bar', 'tristate', 'discrete', 'bullet', 'pie' or 'box'}

\item{width}{Width of the chart - Defaults to 'auto' - May be any valid css width - 1.5em, 20px, etc (using a number without a unit specifier won't do what you want) - This option does nothing for bar and tristate chars (see barWidth)}

\item{height}{Height of the chart - Defaults to 'auto' (line height of the containing tag)}

\item{lineColor}{Used by line and discrete charts to specify the colour of the line drawn as a CSS values string}

\item{fillColor}{Specify the colour used to fill the area under the graph as a CSS value. Set to false to disable fill}

\item{chartRangeMin}{Specify the minimum value to use for the range of Y values of the chart - Defaults to the minimum value supplied}

\item{chartRangeMax}{Specify the maximum value to use for the range of Y values of the chart - Defaults to the maximum value supplied}

\item{composite}{If true then don't erase any existing chart attached to the tag, but draw another chart over the top - Note that width and height are ignored if an existing chart is detected.}

\item{enableTagOptions}{If true then options can be specified as attributes on each tag to be transformed into a sparkline, as well as passed to the sparkline() function. See also tagOptionPrefix}

\item{tagOptionPrefix}{String that each option passed as an attribute on a tag must begin with. Defaults to 'spark'}

\item{tagValuesAtribute}{The name of the tag attribute to fetch values from, if present - Defaults to 'values'}

\item{disableHiddenCheck}{Set to true to disable checking for hidden sparklines. This can be beneficial if you know you'll never attempt to draw a sparkline into a hidden parent element as it avoids a browser reflow for the test, increasing rendering performance.}

\item{spark_options_chart_type}{specific spark_options function for the charttype}

\item{...}{create additional options if required}
}
\value{
javascript call back
}
\description{
Create sparkline options js to include in column call back
}
