% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_helper.R
\name{spark_options_bar}
\alias{spark_options_bar}
\title{Create sparkline options js to include in column call back}
\usage{
spark_options_bar(barColor = NULL, negBarColor = NULL, zeroColor = NULL,
  nullColor = NULL, barWidth = NULL, barSpacing = NULL, zeroAxis = NULL,
  colorMap = NULL, stackedBarColor = NULL, ...)
}
\arguments{
\item{barColor}{CSS colour used for postive values}

\item{negBarColor}{CSS colour used for negative values}

\item{zeroColor}{CSS colour used for values equal to zero}

\item{nullColor}{CSS colour used for values equal to null - By default null values are omitted entirely, but setting this adds a thin marker for the entry - This can be useful if your chart is pretty sparse; perhaps try setting it to a light grey or something equally unobtrusive}

\item{barWidth}{Width of each bar, in pixels (integer)}

\item{barSpacing}{Space between each bar, in pixels (integer)}

\item{zeroAxis}{Centers the y-axis at zero if true (default)}

\item{colorMap}{A range map to map specific values to selected colours. For example if you want all values of -2 to appear yellow, use colorMap: { '-2': '#ff0' }.}

\item{stackedBarColor}{An array of colours to use for stacked bar charts. The first series will use the first value in the array, the second series will use the second, etc}

\item{...}{pass additional options to list}
}
\value{
javascript call back
}
\description{
Create sparkline options js to include in column call back
}
