% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_helper.R
\name{spark_options_box}
\alias{spark_options_box}
\title{Create sparkline options js to include in column call back}
\usage{
spark_options_box(raw = NULL, showOutliers = NULL, outlierIQR = NULL,
  boxLineColor = NULL, boxFillColor = NULL, whiskerColor = NULL,
  outlierLineColor = NULL, outlierFillColor = NULL, spotRadius = NULL,
  medianColor = NULL, target = NULL, targetColor = NULL,
  minValue = NULL, maxValue = NULL, ...)
}
\arguments{
\item{raw}{If set to false (default) then the values supplied are used to caculate the box data points for you. If true then you must pre-calculate the points (see below)}

\item{showOutliers}{If true (default) then outliers (values > 1.5x the IQR) are marked with circles and the whiskers are placed at Q1 and Q3 instead of the least and greatest value}

\item{outlierIQR}{Set the inter-quartile range multipler used to calculate values that qualify as an outlier - Defaults to 1.5}

\item{boxLineColor}{CSS line colour used to outline the box}

\item{boxFillColor}{CSS fill colour used for the box}

\item{whiskerColor}{CSS colour used to draw the whiskers}

\item{outlierLineColor}{CSS colour used to draw the outlier circles}

\item{outlierFillColor}{CSS colour used to fill the outlier circles}

\item{spotRadius}{Radius in pixels to draw the outlier circles}

\item{medianColor}{CSS colour used to draw the median line}

\item{target}{If set to a value, then a small crosshair is drawn at that point to represent a target value}

\item{targetColor}{CSS colour used to draw the target crosshair, if set}

\item{minValue}{If minvalue and maxvalue are set then the scale of the plot is fixed. By default minValue and maxValue are deduced from the values supplied}

\item{maxValue}{See minValue}

\item{...}{pass additional options to list}
}
\value{
javascript call back
}
\description{
Create sparkline options js to include in column call back
}
