% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkline_helper.R
\name{spark_options_pie}
\alias{spark_options_pie}
\title{Create sparkline options js to include in column call back}
\usage{
spark_options_pie(sliceColors = NULL, offset = NULL, borderWidth = NULL,
  borderColor = NULL, ...)
}
\arguments{
\item{sliceColors}{An array of CSS colors to use for pie slices}

\item{offset}{Angle in degrees to offset the first slice - Try -90 or +90}

\item{borderWidth}{Width of the border to draw around the whole pie chart, in pixels. Defaults to 0 (no border)}

\item{borderColor}{CSS color to use to draw the pie border. Defaults to #000}

\item{...}{pass additional options to list}
}
\value{
javascript call back
}
\description{
Create sparkline options js to include in column call back
}
