% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-maintain-order.R
\name{string.to.vect}
\alias{string.to.vect}
\title{String to Vector}
\usage{
string.to.vect(string, sep = "|")
}
\arguments{
\item{string}{Input character string}

\item{sep}{Separator value. Defaults to "|"}
}
\value{
character vector
}
\description{
Splits a string into its constituent parts based on a given separator. 
For example, "A|B" will output "A", "B"
}
\examples{
string <- "A|B"
string.to.vect(string)

string2 <- "A.B"
string.to.vect(string2, sep = ".")

}
