% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_callback.R
\name{styleColorBarStack}
\alias{styleColorBarStack}
\title{Special bar stack function to be used within DT style functions}
\usage{
styleColorBarStack(data, color_1, color_2, angle = 90, min_range, max_range)
}
\arguments{
\item{data}{data}

\item{color_1}{first colour of bar}

\item{color_2}{second colour of bar}

\item{angle}{rotate?}

\item{min_range}{min of chart}

\item{max_range}{of chart}
}
\description{
Special bar stack function to be used within DT style functions
}
