% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{targets_helper}
\alias{targets_helper}
\title{identifies columns from a numeric or character vector - returns js position}
\usage{
targets_helper(x, targets)
}
\arguments{
\item{x}{data}

\item{targets}{character or numeric vector of columns to be found in x}
}
\value{
numeric vector of js positions of columns (i.e expected position - 1)
}
\description{
identifies columns from a numeric or character vector - returns js position
}
