% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{to_js}
\alias{to_js}
\title{check for logical and wrap in '}
\usage{
to_js(x)
}
\arguments{
\item{x}{vector to convert}
}
\value{
character vector
}
\description{
check for logical and wrap in '
}
