% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_sparkline.R
\name{tplSparkline}
\alias{tplSparkline}
\title{Create a sparkline template to be used in a drawback in the DT datatable}
\usage{
tplSparkline(class_name, spark_opts)
}
\arguments{
\item{class_name}{will be prefixed with spark}

\item{spark_opts}{Spark options as javascript (generate with spark_options_)}
}
\value{
single row of javascript to be appended into the fnDrawCallback of the DT
}
\description{
Create a sparkline template to be used in a drawback in the DT datatable
}
