% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-simulation.R
\name{track.time}
\alias{track.time}
\title{Track Time}
\usage{
track.time(i, niter, start, track = TRUE)
}
\arguments{
\item{i}{Iteration number}

\item{niter}{Total number of iterations in loop}

\item{start}{Start time of loop. Use Sys.time() to get start time}

\item{track}{If set to TRUE, reports progress at every 10\% interval}
}
\value{
None. Prints out time differences
}
\description{
Typically used in a loop. Reports progress at every 10% interval.
}
\examples{
start = Sys.time()
for(i in 1:100){
  track.time(i, 100, start)
}
start = Sys.time()
for(i in 1:100){
  track.time(i, 100, start, track =F)
}

}
