% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{vector_to_js_array}
\alias{vector_to_js_array}
\title{converts array to character indicating a js [,] type array}
\usage{
vector_to_js_array(v)
}
\arguments{
\item{v}{vector}
}
\value{
numeric vector of js positions of columns (i.e expected position - 1)
}
\description{
converts array to character indicating a js [,] type array
}
