% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-computation-functions.R
\name{weighted.sum}
\alias{weighted.sum}
\title{Compute Weighted Sum}
\usage{
weighted.sum(x, weights, na.rm = T)
}
\arguments{
\item{x}{Input vector object}

\item{weights}{weight associated to each input value of x}

\item{na.rm}{If set to TRUE, NAs are ignored in computation}
}
\value{
scalar value
}
\description{
Weights are normalised before weighted sum is computed.
}
\examples{
x <- c(1,2,3)
weights <- c(1,1,1) # equal weights
weighted.sum(x, weights)

weights <- c(5, 1, 1) # unequal weights
weight.sum(x, weights)

}
