% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wrap_chr}
\alias{wrap_chr}
\title{wrap characters within a vector with wrap_with}
\usage{
wrap_chr(x, wrap_with = "'")
}
\arguments{
\item{x}{vector to wrap}

\item{wrap_with}{character to wrap stuff}
}
\value{
character vector
}
\description{
wrap characters within a vector with wrap_with
}
