% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dthelper-data-wrangling.R
\name{gather.keep}
\alias{gather.keep}
\title{Row Bind Matching Column Names}
\usage{
gather.keep(
  df,
  keep = NULL,
  var.to.gather = NULL,
  key = "key",
  value = "value"
)
}
\arguments{
\item{df}{data.frame object}

\item{keep}{Variables that shouldn't be gathered}

\item{var.to.gather}{Variable that should be gathered. Defaults to all other variables apart from keep}

\item{key}{Name of key variable after gathering}

\item{value}{Name of value variable after gathering}
}
\value{
data.frame object
}
\description{
row binds a vector to an existing data.frame object (data.frame can be empty).
The vector should be named.
This function ensures that the vector values are added to the column
that shares their respectives names.
}
\examples{
set.seed(100)
df <- data.frame(
      id = 1:10,
      v1_t1 = rnorm(10),
      v1_t2 = rnorm(10),
      v1_t3 = rnorm(10)
      )
gather.keep(df, keep = "id")

}
